# Eliminate stack protection from the runtime.