There is some positive performance effect:
before: Average Ring score: absolute = 177,5160 us +- 0,5352 us, ratio = 8,68 +- 0,10
after:  Average Ring score: absolute = 171,3093 us +- 2,3758 us, ratio = 8,72 +- 0,23